On game over, if the score is a highscore it is added to the highscore dialog. We can show the highscore dialog in the Game menu
svn path=/trunk/playground/games/kapman/; revision=778932

+++ game.cpp
@@ -113,7 +113,7 @@ bool Game::isPaused() const {
 	return m_isPaused;
 }
 
-qreal Game::getScore() const {
+int Game::getScore() const {
 	return m_points;
 }
 int Game::getLives() const {
